Start work on Bonjour support
The idea is that you run FluffyDislay both on your main Mac and on the
one whose display you want to extend the main Mac's desktop to.
FluffyDisplay's menu will contain entries that create virtual displays
suitable for the other machine's display(s), and hopefully I will get
it to work so that when you select one such entry, FLuffyDislay on the
other Mac will be asked to open the Screen Sharing connection to the
main Mac. Possibly even automatically select the virtual display in
question.