Populate the "New for peer" menu
It contains entries for the main display of each peer running
FluffyDisplay in the local network. That might be overdoing it in
local networks with a lot of Macs, of course. We need some settings to
turn off this automatic generation of these menu entries. Or maybe the
setting could be to indicate which is *the* other Mac that we want to
use the physical display of.

Also other cleanup and refactoring.