Request the "other" Mac to open Screen Sharing to the "main" Mac when needed
Pass the request in a TXT record.

This requires that FluffyDisplay is running also on the other Mac, the
one with the display that we want to use as a display of the main Mac.
When that FluffyDisplay sees the TXT record of the main Mac being
updated so that it contains a request for it (the other Mac), it
starts Screen Sharing to the corresponding vnc: URL. This actually
works. But the user has to manually make the window full-screen and
turn on Observer mode. Also, as mentioned in the README.md, there is a
"ghost" cursor movoing on that display even when the real cursor of
the main Mac is on its built-in (or otherwise physically connnected)
display.

Still is missing code to reinstate the TXT record of the main Mac back
to how it was. But maybe that isn't actually necessary. After all, the
main and other Macs aren't in an equal relationship. The idea is not
that the other Mac would ever use a display from the main Mac as its
own. It is always the other way around. The other Mac doesn't really
do anything with the information about the main Mac's displays.