We don't need to advertise the displays of the "main" Mac
The intended use case is that one Mac is the "main" one, the one that
you are using, the one for which you want to use physical displays of
other Macs. It is FluffyDisplay on the main Mac that needs to listen
for TXT records advertising the displays of the other Mac(s). Not the
other way around.

We know that we are running on the "main" Mac when a virtual dislay is
created.