Avoid showing unnecessary submenus
Only show the "Delete" submenu once the user has created a virtual display.

Only show the "New on other Mac" submenu once we have noticed that
FluffyDisplay is running on another Mac.

Don't show anything except "Quit" on an other Mac (once we know we are
running on an "other" Mac).